+++ ts/src/constants.ts
@@ -14,4 +14,6 @@ export const TACTICS_ADV_NAMES_MAP: Record<string, string> = {
   Nefty_Axobubble: 'axobubble',
   Nefty_Unika: 'unika',
   Nefty_Chocomint: 'chocomint',
+  Nefty_Cybertooth: 'cybertooth',
+  Nefty_Wassie: 'wassie',
 };