+++ ts/src/index.ts
@@ -3,4 +3,5 @@ export * from './eggs_factory';
 export * from './dna_schema_reader';
 export * from './interfaces/types';
 export * from './constants';
+export * from './deps/schemas/adventures/latest';
 export * as utils from './utils';